structured-app-ndb/requirements.txt
Flask
mock
google-auth
pytest
google-cloud-ndb
